rust-revlog: take a `u32` in pyo3's `seen_file_size`
The file size are currently limited to u32 everywhere so lets keep things
consistent here.

--HG--
extra : rebase_source : a5bd8d128554231be4863f5c7f244c018a5a92da
extra : topic : u32-utils

+++ rust/hg-pyo3/src/revlog/mod.rs
@@ -497,7 +497,7 @@ impl InnerRevlog {
         })
     }
 
-    fn seen_file_size(slf: &Bound<'_, Self>, size: u64) -> PyResult<PyObject> {
+    fn seen_file_size(slf: &Bound<'_, Self>, size: u32) -> PyResult<PyObject> {
         Self::with_core_read(slf, |_self_ref, irl| {
             irl.seen_file_size(size.try_into().expect("16 bit computer?"));
             Ok(slf.py().None())
